chore: convert @packages/socket tests to vitest and TypeScript
