+++ .circleci/src/pipeline/@pipeline.yml
@@ -1784,7 +1784,7 @@ jobs:
                   source ./scripts/ensure-node.sh
                   yarn lerna run types
             - sanitize-verify-and-store-mocha-results:
-                expectedResultCount: 11
+                expectedResultCount: 10
 
   verify-release-readiness:
     <<: *defaults
